% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{jobScript}
\alias{jobScript}
\title{Resolve the script path of a job}
\usage{
jobScript(job)
}
\arguments{
\item{job}{A \code{BiocJob} object.}
}
\value{
Absolute path to the job's R script.
}
\description{
Resolve the script path of a job
}
