% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{findJobs}
\alias{findJobs}
\title{Discover job specifications in a package}
\usage{
findJobs(pkg = ".", validate = TRUE)
}
\arguments{
\item{pkg}{Path to a package source directory, path to an installed package
directory, or the name of an installed package.}

\item{validate}{Validate each specification (default \code{TRUE}).}
}
\value{
A named list of \code{BiocJob} objects.
}
\description{
Looks for \verb{inst/biocjobs/*.yaml} in a package source directory, or \verb{biocjobs/*.yaml}
in an installed package.
}
