% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{readJob}
\alias{readJob}
\title{Read a single job specification}
\usage{
readJob(path, validate = TRUE)
}
\arguments{
\item{path}{Path to a job YAML file.}

\item{validate}{Validate the specification after reading (default \code{TRUE}).}
}
\value{
A \code{BiocJob} object (a validated list).
}
\description{
Read a single job specification
}
\examples{
yaml <- system.file("examples", "toy", "inst", "biocjobs", "toy-normalize.yaml",
                    package = "BiocJobs")
job <- readJob(yaml)
job
}
